feat(hathorlib): move shielded output parsing to hathorlib
Move shielded output models and serialization from hathor-core to
hathorlib so external services (wallets, tx-mining-service) can parse
transactions with shielded outputs. hathor-core re-exports from
hathorlib to preserve existing import paths.
