+++ hathorlib/hathorlib/vertex_parser.py
@@ -26,10 +26,11 @@ class VertexParser:
     @staticmethod
     def get_supported_headers() -> dict[VertexHeaderId, type[VertexBaseHeader]]:
         """Return a dict of supported headers."""
-        from hathorlib.headers import FeeHeader, NanoHeader, VertexHeaderId
+        from hathorlib.headers import FeeHeader, NanoHeader, ShieldedOutputsHeader, VertexHeaderId
         return {
             VertexHeaderId.NANO_HEADER: NanoHeader,
             VertexHeaderId.FEE_HEADER: FeeHeader,
+            VertexHeaderId.SHIELDED_OUTPUTS_HEADER: ShieldedOutputsHeader,
         }
 
     @staticmethod
